increase memory, change config check to 30 seconds, minor fixes

+++ check-lwaftr-config.sh
@@ -1,22 +1,30 @@
 #!/bin/bash
 PASSWORD=lwaftr1
-CFG=*.cfg
+CONFIG=$1
+touch /tmp/lwaftr-config    # avoid killing snabb the first time
 
 while :
 do
-  IP=`grep fxp0 /u/$CFG|cut -d' ' -f7|cut -d'/' -f1`
+  sleep 30 
+  IP=`grep '172.17.0' /u/$CONFIG |cut -d's' -f 3|cut -d'/' -f1 |sed 's/ //'`
+  if [ "X$IP" == "X" ]; then
+    IP=`grep fxp0 /u/$CONFIG|cut -d' ' -f7|cut -d'/' -f1`
+  fi
+  if [ "X$IP" == "X" ]; then
+    echo "Can't determine IP address of fxp0. Stopping lwaftr check"
+    exit 1
+  fi
 #  echo "vMX fxp0 has IP $IP"
-  sshpass -p lwaftr1 ssh -o StrictHostKeyChecking=no lwaftr@$IP "show conf groups | find lwaftr" > /tmp/lwaftr-config.new
+  sshpass -p lwaftr1 ssh -o StrictHostKeyChecking=no lwaftr@$IP "show conf groups" > /tmp/lwaftr-config.new
   diff /tmp/lwaftr-config.new /tmp/lwaftr-config >/dev/null 2>&1
   if [ $? != 0 ]; then
      rm -f /tmp/lwaftr-xe*.cfg
      if [ -s /tmp/lwaftr-config.new ]; then
-#       echo "new config!"
-       cp /tmp/lwaftr-config.new /tmp/lwaftr-config
-       /parse-lwaftr-config.pl < /tmp/lwaftr-config
-#       echo "restarting snabb"
-       killall snabb
+       # new non empty config. Lets process it!
+       /parse-lwaftr-config.pl < /tmp/lwaftr-config.new
      fi
+     # restart snabb. 
+     cp /tmp/lwaftr-config.new /tmp/lwaftr-config
+     killall snabb
   fi
-  sleep 60 
 done
